refactor: move sticky header logic to separate file for readabiity
